r5761 | Michael.Lueken@noaa.gov | 2009-11-18 09:53:14 -0700 (Wed, 18 Nov 2009) | 2 lines
Trunk: Incorporated changes by Yen-chih @5735 (r5735) to the trunk.  These changes include adding a qc algorithm that if the highest observation in a profile is below 5 km, the rest of observations would be tossed, only affects in regional cases. Also, slightly modify the observation error for the regional cases. Finally, added scripts of regression tests modified for Vapor into the branch.