------------------------------------------------------------------------ r5845 | john.derber@noaa.gov | 2009-11-25 11:29:47 -0700 (Wed, 25 Nov 2009) | 14 lines
Trunk:  Primary modification is an updating of the step size calculations.  The
b and c estimates are moved out of stp* and into stpcalc.  Also the number of function
evaluations is decreased by one in stepsize iteration > 1.  The computions should have
slightly less round-off error.  Because of the changes in the round-off error some small
changes in the results should be expected.


A bug in the printout for those microwave data that use qcssmi for quality control is corrected.
This will result in some changes in the printed output in the radiance diagnostic file.

Also, a couple of minor bugs were removed based on input from the DTC.

John Derber