------------------------------------------------------------------------ r6276 | Michael.Lueken@noaa.gov | 2010-01-12 10:51:07 -0700 (Tue, 12 Jan 2010) | 2 lines
Trunk: Incorporated Manuel Pondeca's and Wan-Shu Wu's changes @6264 (r6264).  For regional and 2DVAR (RTMA), the reference time for the guess files in order to be consistent with the new refernce time used for the observations.  This was done to correct the FGAT results.  For regional, to prepare for 4DVAR, the variable nhr_offset was replaced with min_offset.  It is convenient to count minutes rather than hours, with regional requiring 1.5 hrs (a non-integer value number when expressed in hours).  Everything passed the regression tests.

------------------------------------------------------------------------