------------------------------------------------------------------------ r6575 | john.derber@noaa.gov | 2010-02-04 09:42:01 -0700 (Thu, 04 Feb 2010) | 11 lines
Trunk:  Modify read of data.  These changes fix bugs in the option when
all satellite data (no thinning) is used, improve checking of dates and
existence of data in the setup of the parallel reads, simplify parallel reading of data (
especially read_prepbufr) and general optimization.  Results should reproduce previous code except
note below.  Reads should be a bit faster than before (especially read_prepbufr).

Note a problem with read_amse has been uncovered in this update.  Both the new and the old
do not properly read the amsre data and do not duplicate results when different numbers of
processors are use to read the data.


------------------------------------------------------------------------