------------------------------------------------------------------------ r7707 | Michael.Lueken@noaa.gov | 2010-05-07 13:24:03 -0600 (Fri, 07 May 2010) | 2 lines Changed paths:    M /gsi/trunk/src/Makefile.dependency    M /gsi/trunk/src/control_vectors.f90
Trunk: Ming found an issue with the current trunk running on a Linux PGI compiler.  In pre
wgt_reg.f90, there is a subscript error.  The array nrf_var has the size of 8 but the inde
x nvar_id(l) can be 9 or 10 if sst is analyzed.  Yanqiu had a fix ready to anav_info withi
n control_vectors.f90.  This fix increases the array nrf_var by two if SST data is used as
 a control variable.  This fix has been tested and the results are bitwise identical.  Thi
s should fix the Linux PGI issue as well.