------------------------------------------------------------------------ r9035 | Michael.Lueken@noaa.gov | 2010-07-29 08:37:44 -0600 (Thu, 29 Jul 2010) | 12 lines Changed paths:
The ARW binary regression test scripts were also updated to include an updated ARW binary
test case (as of July 24, 2010 at 12Z).  So long as the l_cloud_analysis flag is set to fa
lse (as is the case in the regression test scripts), then the results are bitwise identica
l.