------------------------------------------------------------------------ r9844 | Michael.Lueken@noaa.gov | 2010-09-15 11:54:57 -0600 (Wed, 15 Sep 2010) | 2 lines Changed paths:    M /gsi/trunk/src/pcgsoi.f90    M /gsi/trunk/src/qnewton.f90    M /gsi/trunk/src/stpcalc.f90
Trunk: Merged derber branch at r9831 to the trunk.  This change brings updates to the pcgs
oi minimization.  Inconsistancies with regard to use of double precision numbers in quad p
recision calculations were fixed in stpcalc and the gnorm(2) calculation is now the averag
e of grady and gradx(n)-gradx(n-1) dot product and the gradx and grady(n)-grady(n-1) dot p
roduct.  The changes affect the results, so there will not be reproducibility between the
trunk and the branches.  One final note, axpy was added to the use control_vectors, only p
ortion of qnewton.f90.  This is necessary for the code to compile on other machines.