------------------------------------------------------------------------ r10021 | Michael.Lueken@noaa.gov | 2010-09-24 09:04:39 -0600 (Fri, 24 Sep 2010) | 12 lines Changed paths:    M /gsi/trunk/src/Makefile.dependency    M /gsi/trunk/src/gsimod.F90    M /gsi/trunk/src/read_tcps.f90    M /gsi/trunk/src/setuptcp.f90    M /gsi/trunk/src/tcv_mod.f90
Trunk: Merged EXP-hybens at r9846 to the trunk.  This change adds flexibility to the TCPS
observation error specification by adding three new namelist options (in &OBSQC):

 * tcp_oberr: base/default ob error (was previously set to 0.75)
 * tcp_innmax: the maximum O-F for which to stop inflating error (this also goes into the
calculation of the "slope" of the inflation. this was previously assumed to be 20mb)
 * tcp_oedelt: the range for which the oberror is to vary as a function of departure from
tcp_oberror to a max value (tcp_oberror+tcp_oedelt). this was previously set to 1.5 (so th
e values ranged from 0.75 to 2.25).

The default values are set to be those that were hardwired for the GFS/GDAS application, i
.e.

 * tcp_oberr=0.75_r_kind
 * tcp_innmax=20.0_r_kind
 * tcp_oedelt=1.5_r_kind