------------------------------------------------------------------------ r10994 | Michael.Lueken@noaa.gov | 2010-11-16 10:56:06 -0700 (Tue, 16 Nov 2010) | 2 lines Changed paths:    M /gsi/trunk/scripts/regression_var.sh    M /gsi/trunk/scripts/run_arw_netcdf_hybens_testcase.sh    M /gsi/trunk/scripts/rungsi_globalprod.sh    M /gsi/trunk/scripts/rungsi_nmmprod.sh
Trunk: Merged lcucu1 branch at r10963 to the trunk.  There were several changes made in th
is revision.  First, in the fix/ directory, the satellite id for unknown GPS RO satellites
, 401, was removed from all convinfo files.  In the scripts/ directory, Lidia noted that u
se_compress was missing from the SETUP namelist option for rungsi_globalprod.sh and rungsi
_nmmprod.sh, so these were added in.  Finally, the main change was to read_gps.f90 in the
src/ directory.  Now, the GSI will skip all profiles which are not listed in the convinfo
file.  Further, the coding was changed so that the ordering of the GPS RO satellites in th
e convinfo files will not affect results.  For further details, please see tickets #139 an
d #142.