------------------------------------------------------------------------ r11739 | Michael.Lueken@noaa.gov | 2011-01-11 06:55:44 -0700 (Tue, 11 Jan 2011) | 2 lines Changed paths:    A /gsi/trunk/scripts/Get_Initial_Files (from /gsi/branches/acollard_runprod/scripts/Get _Initial_Files:11612)    M /gsi/trunk/scripts/rungsi_globalprod.sh
Trunk: Merged acollard_runprod/scripts at r11612 to the trunk.  This update generalizes th
e old rungsi_globalprod.sh script.  All commonly editted components are now grouped at the
 beginning of the script and the script is able to differentiate between the CCS (cirrus a
nd stratus) and Vapor, so users don't need to choose the machine that they are working on.
  Further, after a user chooses an analysis date (adate), the script will first look in th
e com/ directory for the data.  If the data isn't available, the user will be asked to run
 the new script, Get_Initial_Files, which will pull all of the required files from HPSS.
After this, the script will run to completion.

------------------------------------------------------------------------