----------------------------------------------------------------------- r12717 | Michael.Lueken@noaa.gov | 2011-03-10 12:41:26 -0700 (Thu, 10 Mar 2011) | 2 lines Changed paths:    M /gsi/trunk/scripts/rungsi_globalprod.sh
Trunk: Merged acollard_prod at r12708 to the trunk.  This update corrects an issue with th
e rungsi_globalprod.sh script.  If the script was run at T382 resolution after the date of
 the operational resolution increase, the script would fail due to assuming that the opera
tional resolution was still T382.  Changes were made to check the resolution of the guess
file.  The path definitions for the executables were also moved to the top of the script.

------------------------------------------------------------------------