------------------------------------------------------------------------ r13411 | Michael.Lueken@noaa.gov | 2011-04-18 10:47:09 -0600 (Mon, 18 Apr 2011) | 2 lines Changed paths:    M /gsi/trunk/src/read_radar.f90    M /gsi/trunk/src/setuprw.f90
Trunk: Merged sliu branch at r13152 to the trunk.  This update is to the radial wind data.
  First, the observation subtype is being used to discriminate between level-2, level-2.5,
 and level-3 observations.  Further, due to the redundant radial wind information in the r
egional analysis for the CONUS domain, level-2.5 radial winds will not be used in the CONU
S domain.  This data will be used for Alaska due to no level-2 data avaialble.  Finally, t
he observation error has been increased for level-3 data in the CONUS domain to reduce the
 impact of this data on the analysis.  Only two routines have been modified, read_radar.f9
0 and setuprw.f90.

------------------------------------------------------------------------