svn merge -r18041:18083 https://svnemc.ncep.noaa.gov/projects/gsi/trunk/src ------------------------------------------------------------------------ r18083 | Michael.Lueken@noaa.gov | 2012-03-05 06:54:16 -0700 (Mon, 05 Mar 2012) | 10 lines Changed paths:    M /gsi/trunk    M /gsi/trunk/fix    M /gsi/trunk/scripts    M /gsi/trunk/scripts/regression_var.sh    M /gsi/trunk/src    M /gsi/trunk/src/constants.f90    M /gsi/trunk/src/gesinfo.f90    M /gsi/trunk/src/gsimod.F90    M /gsi/trunk/src/makefile_DTC    M /gsi/trunk/src/ncepnems_io.f90    M /gsi/trunk/src/obsmod.F90    M /gsi/trunk/src/read_aerosol.f90    M /gsi/trunk/src/read_files.f90    M /gsi/trunk/src/read_prepbufr.f90    M /gsi/trunk/src/read_satwnd.f90    M /gsi/trunk/src/setupq.f90    M /gsi/trunk/src/setupt.f90    M /gsi/trunk/src/setupw.f90    M /gsi/trunk/util
Trunk: Merged wwu_extsonde_test at 18064 to the trunk.  The goal was to use the significant levels of the radiosonde to help reduce non-mete
orological profiles (i.e., super saturation) in the analysis, or reducing the possibility of analysis profiles being further away from the o
bservation than the first guess. With 3dvar, the innovations were spread in the vertical without considering the first guess structure. This
 could produce non-realistic profiles around planetary boundary layer. For example, if the model PBL was lower than the observed, the q obs
near boundary layer top that help to increase the humidity above model's PBL might force the humidity profile below to pass the saturation v
alue. The work was to constraint the solution with bogus observations by using significant level information.
The condition to generate bogus data were

 1) one of the two obs that the grid point located between was significant
 2) both upper and lower obs were of good quality (used)
 3) both obs are from the same station.
 4) above first level from the ground.

The bogus data were created on the model grid where no observations were taken in the near by grid interval. The bogus data were generated f
or temperature, wind and humidity. The code could be activated by setting the parameter .ext_sonde. to .true. The significant observations w
ere identified by the the flag .CAT. in prepbufr file. CAT for significant winds was 3 and was 2 for mass data. CAT for tropopause was 5 whi
ch was considered as significant also.