***************  Merge from HWRF branch **************** Merge r967
Modified:
   branches/HWRF/src/libs/nemsio/Makefile

Log:
Fixed a bug in a makefile that was preventing the GSI from compiling on CCS
Stratus and Cirrus.  The nemsio makefile was trying to use the Fortran 77
compiler ($FC) to compile Fortran 90 (*.f90) files.  The GNU, PGI and Intel
compilers are smart enough to handle that, but the IBM compiler is not.  I
changed the makefile to use the Fortran 90 compiler ($F90).