##################Runtime Bug Fix ################################## Rewrote write statements for logging error messages in mprint_ routines. When a namelist error logged an error message, the write to standard error and out failed.
Replaced:
  write(lu,'(3a)',advance='no') &
    trim(myid_(who)),'(): ',trim(what)

with
  ltmp= trim(ltmp) // '(): ' // trim(what)
  write(lu,'(a)',advance='no') trim(ltmp)


The runtime error occurs with two different versions of two different compilers (pgi and intel). Discovered the error
when I introduced an error in the namelist. the run triggered the error call after the read gridopts. 

 835   read(11,gridopts,iostat=ios)
 836         if(ios/=0) call die(myname_,'read(gridopts)',ios)

The call to die, calls the mprint_ routines to print the error message to standard error and standard out.