------------------------------------------------------------------------ r29302 | Michael.Lueken@noaa.gov | 2013-06-19 11:03:45 -0600 (Wed, 19 Jun 2013) | 9 lines Changed paths:    M /gsi/trunk    M /gsi/trunk/fix    M /gsi/trunk/scripts    M /gsi/trunk/scripts/regression_var.sh    M /gsi/trunk/src    M /gsi/trunk/src/constants.f90    M /gsi/trunk/src/gesinfo.f90    M /gsi/trunk/src/gsimod.F90    M /gsi/trunk/src/ncepnems_io.f90    M /gsi/trunk/src/read_aerosol.f90    M /gsi/trunk/src/read_files.f90    M /gsi/trunk/src/read_prepbufr.f90    M /gsi/trunk/src/setupq.f90    M /gsi/trunk/src/setupt.f90    M /gsi/trunk/src/setupw.f90    M /gsi/trunk/util
Trunk: Reintegration merged wwu_sonde at r29276 to the trunk. It was found that current code generated bogus observations between a significant level report and one which was located much higher than the next observation. The reason was the radiosonde reports drift out of a sub-domain and drift back into the sub-domain. The bogus data would be inaccurate and the inconsistent bogus data between sub-domains could cause problems in minimization. The attempt to fix the problem in current coding was failed. It was found that while the problem of the radiosonde leaving the sub-domain can be fixed, it was more complicated to produce consistent bogus when the RAOB drifted back into the sob-domain. It was decided to redo the option in subroutine READ_PREPBUFR, when the data were read in.

* src:
 * setupw.f90 - moved ext_sonde to read_prepbufr
 * setupt.f90 - same
 * setupq.f90 - same
 * read_prepbufr.f90 - same
 * gsimod.F90 - write ext_sonde output to standard out