r61647 | Michael.Lueken@noaa.gov | 2015-09-14 16:53:17 +0000 (Mon, 14 Sep 2015) | 6 lines
Trunk: Merged EXP-Theia_transition branch at r61641 to the trunk. Since Zeus has reached it's EOL, it is necessary to update the regression tests to use Theia only (not Zeus or Theia) and update the directory paths to all point to Theia's scratch3 or scratch4.
 * Made necessary changes to multi_regression.sh, regression_var.sh, and sub_zeus to point to Theia directories.
 * Removed $machine = Zeus entries from all regression test scripts.
 * Made modifications to regression_test.sh and regression_test_enkf.sh to remove the working directories once the regression tests for a specific configuration is complete.
With these changes, the regression tests have been running smoothly on Theia and there have been no further dumps to the home directory (and subsequent emails from Theia).