Sync EMC trunk 65747, 66575,67694,68107 to community trunk.
------------------------------------------------------------------------
r68107 | Michael.Lueken@noaa.gov | 2016-01-25 18:06:52 +0000 (Mon, 25 Jan 2016) | 2 lines

Trunk: Merged esafford_RadMon_540 branch at r67990 to the trunk.  The gdas_radmon_base.tar contains the historic penalty bound values for each instrument/channel, and is used to identify possible invalid penalty values in the assimilated radiance data. This file has been updated for the pr4devb parallel, and that version of the file needs to become the new default version within the !RadMon package.

------------------------------------------------------------------------
r67694 | Michael.Lueken@noaa.gov | 2016-01-15 19:07:20 +0000 (Fri, 15 Jan 2016) | 22 lines

Trunk: Merged minim branch at r67522 to the trunk.  The global pre-implementation parallel has been showing occasional issues in the minimization. Many resets and occasional early stopping. Looking at the output indicates that much of the issue occurs because of competition between the supersaturation constraint and the other moisture terms. The minimization would work properly when the supersaturation constraint was turned off.

It does not appear that there is any problem with the supersaturation constraint, but rather the issue is the background error covariance for the q field. By putting a simple limit on the input moisture (0<=q<=1.2*qsat) in the ensembles, the problem went away. However this limit has some problems, because if all of the ensembles are greater than supersaturation, or less than zero, the variance becomes zero and no change is allowed in the analysis at this location.

A second solution was tried. In this, the difference of the mean ensemble field at each analysis time (e.g., -3,-2,-1,0,1,2,3hr) was limited to be between -100% and 100%. This resulted in a clean minimization and much improved fits to obs for all variables for the cases the minimization failed.

Russ ran this version on other cases where there were not minimization issues. The results did change, but no significant differences were noted.

Those points that were adjusted were printed out. They showed values that input values that were often > 200% of saturation (sometimes much greater). This seems very unphysical and may mean we need to revisit how we are perturbing the ensembles and/or when we are writing the forecasts. Currently we are writing the output immediately after the perturbation. So if the q field is perturbed to greater than saturation, the physics does not have a chance to precipitate the additional moisture.

Changes necessary to limit the difference from the mean to +-100% RH are put into branch minim.

Additionally, the aircraft bias correction could play a role in the "Resets" occurrences.

After reviewing the GSI routines, it was discovered that the preconditioning for the aircraft bias correction was only applied to the cases when obs_count>3, but not to the cases when obs_count<=3 as it was undesirable to use rstats_t information from these very few observations in the preconditioning formula when the code was originally written. It is now suspected that this may cause a problem for an observation from preconditioned in a cycle to un-preconditioned to the next cycle, so it was suggested to apply the preconditioning formula to the cases with obs_count<3 as well.

Only two routines were modified in this update:

 * src:
   * get_gefs_ensperts_dualres.f90
   * berror.f90

------------------------------------------------------------------------
r66575 | Michael.Lueken@noaa.gov | 2015-12-15 11:27:42 +0000 (Tue, 15 Dec 2015) | 37 lines

Trunk: Merged gsi_nsst_update branch at r66047 to the trunk.  This update is an inclusion of two tickets and a regional bug fix:

#453:

A few NSST updates are addressed in this ticket:

 * Move the interpolation of Tf (foundation temperature) analysis increment (from the analysis grids to ensemble grids) to recenter step of EnKF.
 * Modify to enable to provide the SST, which is used at the forecasting mode, in the form of vertical mean or at a specific depth, based the T-Profile produced by NSST Model (NSSTM).
 * Modify to read nst file in the same way as for sfc and sig files, which was update at the end of 2014 by John.
 * Deallocate the nst full resolution arrays earlier to save memory usage.
 * Modify the real variables read from sfc and nst files to be single precision (from double precision) to save memory usage.
 * Modify to make it work when nst_gsi = 0 and nsstbufr data file exists.

#522:

Porting the GSI and EnKF onto the new Cray machine. This ticket will be updated with pertinent information relating to necessary environment variables and when updates have been made to the code.

Bug Fix:

There was a operational NDAS failure on 20151126 12 z. It was caused by a radiosonde report that had more than 255 levels. A bug in subroutine sonde_ext in file read_prepbufr.f90 was found. The check to prevent array out of bound was in place but after printing out an error message the program was stopped.

    if(ll>255)then

        write(6,*)'error in SONDE_EXT levs > 255'
        STOP

    endif

A fix to return to the calling routine which would allow the analysis to continue was included in this branch.

    if(ll>255)then

        write(6,*)'error in SONDE_EXT levs > 255'
        RETURN

    endif

------------------------------------------------------------------------
r65747 | Michael.Lueken@noaa.gov | 2015-11-27 18:09:43 +0000 (Fri, 27 Nov 2015) | 2 lines

Trunk: Merged rtma_q1fy16 branch at r65700 to the trunk.  Testing of the RTMA for the Q2FY16 implementation has shown an issue with the new non-linear quality control (njqc) for the surface winds.  Turning off the njqc for the surface wind analysis corrected this issue.  This update makes the necessary changes to the prepobs_errtable_uv.optm and nlqc_b_uv.rtma.optm of surface wind analysis to mirror the Q2FY16 RTMA system.